Fix clippy and enable clippy on CI

+++ src/data/arena.rs
@@ -269,7 +269,7 @@ impl<T> Arena<T> {
                     self.free_list_head = next_free;
                     self.len += 1;
                     Some(Index {
-                        index: i as u32,
+                        index: i,
                         generation: self.generation,
                     })
                 }
